+++ src/projects/components/project-card/project-card.styles.ts
